Conflicts:
	fortscale/fortscale-collection/src/main/java/fortscale/collection/jobs/activity/UserActivityBaseHandler.java
	fortscale/fortscale-collection/src/main/java/fortscale/collection/jobs/activity/UserActivityHandlerFactory.java
	fortscale/fortscale-collection/src/main/java/fortscale/collection/jobs/activity/UserActivityJob.java
	fortscale/fortscale-collection/src/main/java/fortscale/collection/jobs/activity/UserActivitySourceDevicesHandler.java
	fortscale/fortscale-collection/src/main/java/fortscale/collection/jobs/activity/UserActivityType.java
	fortscale/fortscale-collection/src/main/java/fortscale/collection/services/UserActivityLocationConfigurationService.java
	fortscale/fortscale-collection/src/main/java/fortscale/collection/services/UserActivityNetworkAuthenticationConfigurationService.java
	fortscale/fortscale-common/src/main/java/fortscale/common/datastructures/UserActivityEntryHashMap.java
	fortscale/fortscale-domain/src/main/java/fortscale/domain/core/activities/UserActivityDocument.java
	fortscale/fortscale-domain/src/main/java/fortscale/domain/core/activities/UserActivityDocumentFactory.java
	fortscale/fortscale-domain/src/main/java/fortscale/domain/core/activities/UserActivityNetworkAuthenticationDocument.java
	fortscale/fortscale-domain/src/main/java/fortscale/domain/core/dao/UserActivityRepositoryImpl.java
	fortscale/fortscale-services/src/main/java/fortscale/services/UserActivityService.java
	fortscale/fortscale-services/src/main/java/fortscale/services/impl/UserActivityServiceImpl.java
	fortscale/fortscale-webapp/src/main/java/fortscale/web/rest/ApiOrganizationActivityController.java
	fortscale/fortscale-webapp/src/main/java/fortscale/web/rest/ApiUserActivityController.java
	fortscale/fortscale-webapp/src/main/java/fortscale/web/rest/entities/activity/UserActivityData.java

+++ fortscale/fortscale-collection/src/main/java/fortscale/collection/jobs/activity/UserActivitySourceMachineHandler.java
@@ -26,7 +26,8 @@ public class UserActivitySourceMachineHandler extends UserActivityBaseHandler {
 	@Override
 	protected List<String> getRelevantFields(String dataSource) throws IllegalArgumentException {
 		final String dataSourceLowerCase = dataSource.toLowerCase();
-		UserActivitySourceMachineConfigurationService.UserActivityDataSourceConfiguration conf = userActivitySourceMachineConfigurationService.getActivityDataSourceConfigurationMap().get(dataSource);
+		UserActivitySourceMachineConfigurationService.UserActivityDataSourceConfiguration conf =
+				userActivitySourceMachineConfigurationService.getActivityDataSourceConfigurationMap().get(dataSource);
 		if (conf != null) {
 			return new ArrayList<>(Collections.singletonList(conf.getFeatureName()));
 		}
@@ -42,7 +43,8 @@ public class UserActivitySourceMachineHandler extends UserActivityBaseHandler {
 			return (GenericHistogram) ((Feature) objectToConvert).getValue();
 		}
 		else {
-			final String errorMessage = String.format("Can't convert %s object of class %s", objectToConvert, objectToConvert.getClass());
+			final String errorMessage = String.format("Can't convert %s object of class %s", objectToConvert,
+					objectToConvert.getClass());
 			logger.error(errorMessage);
 			throw new RuntimeException(errorMessage);
 		}
@@ -59,7 +61,8 @@ public class UserActivitySourceMachineHandler extends UserActivityBaseHandler {
 	}
 
 	@Override
-	protected void updateAdditionalActivitySpecificDocumentInDatabase(List<String> dataSources, long currBucketStartTime, long currBucketEndTime, Map<String, Integer> additionalActivityHistogram) {
+	protected void updateAdditionalActivitySpecificDocumentInDatabase(List<String> dataSources,
+			long currBucketStartTime, long currBucketEndTime, Map<String, Double> additionalActivityHistogram) {
 		//do nothing
 	}
 
@@ -82,4 +85,5 @@ public class UserActivitySourceMachineHandler extends UserActivityBaseHandler {
 	protected UserActivityConfigurationService getUserActivityConfigurationService() {
 		return userActivitySourceMachineConfigurationService;
 	}
+
 }
